refactor: using index.js for page components
