+++ app/components/global/index.js
@@ -1,0 +1,5 @@
+export { default as AppBar } from './AppBar'
+export {
+  default as AuthenticatedComponent,
+} from './AuthenticatedComponent'
+export { default as Layout } from './Layout'
