--- app/components/index.js
@@ -1,6 +1,0 @@
-export { default as AppBar } from './global/AppBar'
-export {
-  default as AuthenticatedComponent,
-} from './global/AuthenticatedComponent'
-export { default as Dashboard } from './pages/dashboard/Dashboard'
-export { default as LoginPage } from './pages/login/LoginPage'
