+++ app/components/pages/dashboard/index.js
