+++ app/components/pages/manuscript/index.js
