+++ app/routes.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 
-import { AuthenticatedComponent, LoginPage } from './components'
-import Layout from './components/global/Layout'
-import SubmissionPage from './components/pages/submission-wizard/SubmissionPage'
-import ManuscriptPage from './components/pages/manuscript/ManuscriptPage'
-import DashboardPage from './components/pages/dashboard/DashboardPage'
+import { AuthenticatedComponent, Layout } from './components/global'
+import LoginPage from './components/pages/login'
+import SubmissionPage from './components/pages/submission-wizard'
+import ManuscriptPage from './components/pages/manuscript'
+import DashboardPage from './components/pages/dashboard'
 
 const Routes = () => (
   <Switch>
